backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-api-python-client==2.154.0
httpx<0.28
python-multipart==0.0.20
anthropic==0.69.0
descope==1.0.0
